use Google Oauth Logout, combine in one component
