+++ client/src/components/pages/Main.jsx
@@ -1,11 +1,10 @@
 import React, { useEffect } from 'react';
 import { Route, Switch } from 'react-router-dom';
 import { connect } from 'react-redux';
-import { MDBIcon, MDBBtn } from 'mdbreact';
 import Dashboard from './Dashboard';
 import MyProfile from './MyProfile';
 import MyHive from './MyHive';
-import GoolgeLogin from '../GoolgeLogin';
+import GoolgeOAuth from '../GoolgeOAuth';
 import Spinner from '../Spinner';
 import '../../styles/pages/Main.scss'
 import api from '../../api';
@@ -13,7 +12,6 @@ import profilePicPlaceholderUrl from '../../media/beesaberlogo.png'
 
 function Main(props) {
     const { dispatch, userdata, fetchingData } = props;
-
     let handleLogout = () => {
         api.logout();
         let userdata = { username: null, profilePic: null }
@@ -44,13 +42,8 @@ function Main(props) {
                     {api.isLoggedIn() && (fetchingData.status) && <Spinner text={fetchingData.statusText} />}
                 </div>
                 <div className="headerpart" id="header-right">
-                    {/* TODO: Move Google Login right */}
-                    {!api.isLoggedIn() && <GoolgeLogin id="googlelogin"/>} 
-                    {api.isLoggedIn() && 
-                        <MDBBtn onClick={handleLogout} size="sm" color="danger">
-                            Logout
-                            <MDBIcon icon="sign-out-alt" className="ml-1" />
-                        </MDBBtn>}
+                    {!api.isLoggedIn() && <GoolgeOAuth id="googlelogin" type="login"/>} 
+                    {api.isLoggedIn() &&  <GoolgeOAuth id="googlelogin" type="logout" />} 
                 </div>
             </div>
             <Switch>
